third_party/blink/web_tests/platform/win/external/wpt/html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/fieldset-content-before-legend-expected.txt

This is a testharness.js-based test.
FAIL fieldset content before legend assert_equals: expected 53 but got 74
Harness: the test ran to completion.
